requirements.txt
torch>=2.1
torchrl>=0.4
tensordict>=0.4

gymnasium>=0.29
numpy>=1.23

hydra-core>=1.3
omegaconf>=2.3

opencv-python>=4.8
Pillow>=10.0

imageio>=2.31
imageio-ffmpeg>=0.4
tqdm>=4.66
matplotlib>=3.7